requirements.txt
langchain
openai
faiss-cpu 
pypdf 
streamlit 
python-dotenv
